Use uuid::Uuid in public API rather than defining our own UUID type.
A 16-bit UUID isn't a different kind of UUID, it's just a short encoding for
a particular subset of UUIDs. This can be handled by an extension trait and
a couple of helper functions.

These were copied from the bluez_async crate, but I wrote them and I'm happy
for them to be under the btleplug license.